requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.24.1
python-dotenv==0.20.0
azure-cosmos==4.9.0
